refactor(client): moves component declarations, styles and tests to a single folder per component
