+++ client/src/components/MainComponent/styles.css
