--- client/src/components/TextInput.js
@@ -1,7 +1,0 @@
-import '../component-styles/TextInput.css';
-
-const TextInput = props => (
-  <input className="text-input" autoComplete="off" {...props}></input>
-);
-
-export default TextInput;
