+++ client/src/components/stages/Drawing/styles.css
