+++ client/src/components/stages/Home/home.test.js
@@ -1,6 +1,6 @@
 import {fireEvent, render, screen} from '@testing-library/react';
 import React from 'react';
-import Home from '../HomeComponent.js';
+import Home from '.';
 
 describe('Home', () => {
   test('Renders error message if present', () => {
