+++ client/src/components/stages/Presentation/presentation.test.js
@@ -1,6 +1,6 @@
 import {render, screen} from '@testing-library/react';
 import React from 'react';
-import Presentation from '../PresentationComponent.js';
+import Presentation from '.';
 
 describe('Presentation', () => {
   test('Renders the', () => {
