+++ client/src/components/stages/Problem/problem.test.js
@@ -1,6 +1,6 @@
 import {fireEvent, render, screen} from '@testing-library/react';
 import React from 'react';
-import Problem from '../ProblemComponent.js';
+import Problem from '.';
 
 describe('Problem', () => {
   test('Renders sample problem', () => {
