+++ client/src/components/stages/Results/styles.css
