+++ client/src/components/stages/Vote/styles.css
