+++ client/src/components/ui/TextInput/index.js
@@ -1,0 +1,7 @@
+import styles from './styles.module.css';
+
+const TextInput = props => (
+  <input className={styles.input} autoComplete="off" {...props}></input>
+);
+
+export default TextInput;
