+++ client/src/components/ui/TextInput/styles.module.css
@@ -1,4 +1,4 @@
-.text-input {
+.input {
   color: var(--color-heading);
   padding: 1rem 2rem;
   border: none;
